requirements.txt
annotated-types==0.6.0
blinker==1.7.0
click==8.1.7
dnspython==2.4.2
Flask==3.0.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pydantic==2.5.3
pydantic_core==2.14.6
pyee==11.1.0
PyJWT==2.8.0
pymongo==4.6.1
python-dotenv==1.0.0
python-ffmpeg==2.0.9
typing_extensions==4.9.0
Werkzeug==3.0.1